move plotting functions to a diff file
